Seperate out detectLanguage into its own ts module
